Boiler plate code for PokeSurface class
Not sure if  I will keep this class but thus far it proves usefull. The only issue is when the position of the rect isnt straight forward. Added FPS Counter aswell. Split onInit into two function, onInit and onSLDInit. Add a few outputs to get a better sense of whats happening in the program.