Fully integrated lessons from AppState management
Took a few days but I've finally integrated the lessons from AppState management. Need to flesh it out more to function for the needs of my project but overall its there. Should make creating other windows/activity easier from here on out. The next step is either to bring in the list from the other project as is or create it from scratch given the newly provided functionality of Activityand Event class.