Minor changes to fix image scaling
Move Img_Init to the top of initSDL in Pokedex.cc. Attempting to add fade effect to intro in introActivity, stil needs work. Moved loading image to Onloop of ListActivity, assuming I will need this in the future to make sure changes render after init. Add a check in setActivitystate of activityManager class to check activity is not none before proceeding. changed BlitSurface to BlitScaled in Surface class, needed for scaling images.