Create PokeActivity_Moves
Changed a lot. added small queries in sqlstatements.cpp and a new one for moves. Created new variable in PokedexDB to keep track of regionGroupID. Might use it might not. Made slight change to query of getPokemonNameandID. Now it gets pokemon by generation, meaning if we get the assets for all the gens we can include those gens them to the project.