Fix: Fixed queries causing slowdown and errors
Queries were changed to filter on gneration. For some reason, probably a line of code I lost track of, the program slowed down when switching screens and some data was not retrieved correctly (evolution chain would only return the first pokemon ever selected). Revered the queries back to what they use to be for the most part.