Feat: Created PokedexActivitySetting, allows user to select setting options
This was a lot of effort for me to get right. And this is not right. Currently it is a working settings page so I will stick with it. But it writes to the file every time the user changes the option even befor the select to exit the screen. Could be an expected behaviour on my part. If they press B it should write even if they navigated the options. Trying to start the options at a specific index means I have to figure out where in the current settingOptions the users selected setting is to find the index.

 Very sloppy, dont understand, but its a starting point.