Fix: Including art work, no longer relying on curl to download.
This might opne the door for a certain company to target this project, but this will allow all miyoo devices to use the app, even if it doesnt have wifi. We can also drop the curl dependency. Perhaps I should make a seperate folder 'res' that is seperate from Source and App folders instead of uploading the contents of res three times across the different versions of the project source and executable