Chore: Move res folder to App/Retrodex/res
Again, only keeping one copy of this folder for simplicity and project size management. Will add instructions for devs to create a symlink from  to the  or . Think I will do something similar with code, just have one copy and leave it up to the dev how to set up their local env.