Chore: Add code to target evolution method using local language id
Was able to arrive at a query that uses language id to get the evolution method but the only language that's really supported is language id 9 ( English ). This means that we can't support displaying the evo methods in other languages.. I belive this is the 3rd time I encounter lack of language support for certain data, this also effects routes/location only English is supported. Addressing issue#9