Chore #11:Ignoring window_dev folder
Adding window_dev folder/env to gitignore. Simplest way to consolidate the two, just get rid of one :D The other option would be to try and create the necessary environment in the union-toolchain to cross compile for windows env. I tried doing this for a bit with MinGW but wasnt able to succesfully build the SDL2 lib using mksdl2.sh. The previous commit should have the files used for that if anyone in the future wants to pick up the charge. Going to distribute executable for miyoo mini device and soure code so folks can inspect, contribute, or fork. Besides my personal pref to dev on windos and ease of testing code changes on local machine, there's not much point in sharing windows dev env configs. Leaving it up to contributors how they setup their local dev env