Clean: Change sqlite include folder stucture
Removed unnecesarry sqlite.o and duplicate sqlite folder. This project uses the sqlite amalgamation. Which means all the necessary code to include the library during compiling is in the sqlite3.h file. The complete implementation of the library is in the .c file. These are amalgamation files, not source code for sqlite3 itself. There is no need to compile sqlite3 from source code when using the amalgamation file. Nor is there a need for the .o or .so files. sqlite3.h is added to -Include directory and can easily be refrenced in preprocessor directives. The sqlite library code is added during preprocessing to the files that call it like other pre-process directives.